Moved bulit in functions to "src/builtins.c"
